requirements.txt
sgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
tzdata==2022.7
django-admin-interface==0.25.0
djangorestframework==3.14.0
markdown==3.4.3